chore(side_shift): move common utils
